requirements.txt
asgiref==3.5.2
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-yaml==2.0.0
Markdown==3.3.7
mysqlclient==2.1.0
Pygments==2.12.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tzdata==2022.1
